requirements.txt
asgiref==3.11.0
attrs==25.4.0
boto3==1.35.90
botocore==1.35.99
certifi==2025.11.12
charset-normalizer==3.4.4
dj-database-url==2.3.0
Django==5.2.8
django-ckeditor-5==0.2.15
django-cors-headers==4.9.0
django-filter==24.3
django-js-asset==3.1.2
django-storages==1.14.4
django-unfold==0.42.0
djangorestframework==3.16.1
drf-spectacular==0.27.2
drf-yasg==1.21.11
easy-thumbnails==2.10
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rpds-py==0.29.0
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.8.2
